requirements.txt
python-docx>=0.8.11
PyPDF2>=3.0.0
reportlab>=4.0.4